feat: Добавлен репозиторий для результатов. Дописан контроллер и сервис результатов. Добавлены exceptions результатов.

+++ src/main/java/com/requestdesign/testingservice/controller/result/ResultControllerInterface.java
@@ -6,11 +6,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 
 public interface ResultControllerInterface {
     @GetMapping("/result/{resultId}")
-    public ResponseEntity getResultById(@PathVariable Integer resultId);
+    public ResponseEntity getResultById(@PathVariable Long resultId);
 
     @GetMapping("/result/{resultId}/task/{taskId}")
-    public ResponseEntity getTaskFromResultById(@PathVariable Integer resultId, @PathVariable Integer taskId);
+    public ResponseEntity getTaskFromResultById(@PathVariable Long resultId, @PathVariable Long taskId);
 
     @GetMapping("/result/{resultId}/question/{questionId}")
-    public ResponseEntity getQuestionAnswerFromResultById(@PathVariable Integer resultId, @PathVariable Integer questionId);
+    public ResponseEntity getQuestionAnswerFromResultById(@PathVariable Long resultId, @PathVariable Long questionId);
 }
